Bugs/fix invalid lease price since mainnet3 (#49)
* Use DecCoin for bid/lease price

* Bump version to 1.2.6

+++ api/src/db/schema.ts
@@ -280,7 +280,7 @@ Bid.init(
     gseq: { type: DataTypes.INTEGER, allowNull: false },
     oseq: { type: DataTypes.INTEGER, allowNull: false },
     provider: { type: DataTypes.STRING, allowNull: false },
-    price: { type: DataTypes.INTEGER, allowNull: false },
+    price: { type: DataTypes.DECIMAL, allowNull: false },
     createdHeight: { type: DataTypes.INTEGER, allowNull: false }
   },
   {
